refactored UI components + loading
