+++ src/Windows/SaveWindow.cs
@@ -33,8 +33,10 @@ namespace linerider.Windows
 {
     class SaveWindow : Window
     {
-        public SaveWindow(Gwen.Controls.ControlBase parent, GLWindow game) : base(parent, "Save Track")
+        private GLWindow game;
+        public SaveWindow(Gwen.Controls.ControlBase parent, GLWindow glgame) : base(parent, "Save Track")
         {
+            game = glgame;
             game.Track.Stop();
             MakeModal(true);
 
@@ -66,7 +68,74 @@ namespace linerider.Windows
                 cb.AddItem(trackname);
             }
             cb.SelectByText(game.Track.Name);
+            var btn = new Button(bottom) { Name = "savebtn" };
+            btn.Width = 50;
+            btn.Text = "Save";
+            btn.Dock = Pos.Right;
+            btn.Clicked += savebtn_Clicked;
+            DisableResizing();
             game.Cursor = MouseCursor.Default;
+        }
+        private void savebtn_Clicked(ControlBase sender, ClickedEventArgs arguments)
+        {
+            var window = sender.Parent.Parent as WindowControl;
+            if (window == null)
+                throw new Exception("Invalid window data");
+            if (window.UserData != null)
+            {
+                var tb = (TextBox)window.FindChildByName("tb", true);
+                var saveindex = 0;
+                var txt = (string)window.UserData;
+                if (txt == "<create new track>")
+                {
+                    txt = tb.Text;
+                    if (txt.Length == 0)
+                        return;
+                }
+                if (
+                    Directory.Exists(Program.CurrentDirectory + "Tracks" + Path.DirectorySeparatorChar + txt +
+                                     Path.DirectorySeparatorChar))
+                {
+                    var trackfiles =
+                        TrackLoader.EnumerateTRKFiles(Program.CurrentDirectory + "Tracks" + Path.DirectorySeparatorChar +
+                                                      txt);
+                    for (var i = 0; i < trackfiles.Length; i++)
+                    {
+                        var s = Path.GetFileNameWithoutExtension(trackfiles[i]);
+                        var index = s.IndexOf(" ", StringComparison.Ordinal);
+                        if (index != -1)
+                        {
+                            s = s.Remove(index);
+                        }
+                        if (int.TryParse(s, out saveindex))
+                        {
+                            break;
+                        }
+                    }
+                }
+                var invalidchars = Path.GetInvalidFileNameChars();
+                for (var i = 0; i < txt.Length; i++)
+                {
+                    if (invalidchars.Contains(txt[i]))
+                    {
+                        sender.SetToolTipText("Attempted to save with an invalid name");
+                        return;
+                    }
+                }
+                game.Track.Name = txt;
+                saveindex++;
+                var save = saveindex + " " + tb.Text;
+                try
+                {
+                    game.Track.Save(save, game.CurrentSong);
+                }
+                catch
+                {
+                    sender.SetToolTipText("An error occured trying to save");
+                    return;
+                }
+            }
+            window.Close();
         }
     }
 }
